requirements.txt
discord.py==1.4.1
neispy==3.1.0
aiomysql==0.0.20
jishaku==1.19.1.200
aiohttp==3.6.2
tortoise-orm==0.16.14